Add form id as criteria in storage prefix
This allows us to handle forms which have two inputs of the same name.

In Rails, if you do a multiselect form element, it adds a hidden input with an empty value, to make sure that you always submit something back to the server, such that you can select the empty set. This hidden input has the same name as the select, but is distinguishable by id, so we want to also use that to store and restore values.