Added onBeforeSave hook
I think was usefull if you want to save all fields and control when to save (with a button, for example). 
Saving is triggered automatically on protect(), so if you want to save data manually, you will need to initialize onBeforeSave to return false. Before manual save, onBeforeSave should be set to return true.